requirements.txt
pandas
polars
pyarrow
openpyxl
dagster
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
plotly
plotly-express